Fix compilation warning in TclListObj - printf format mismatch

+++ generic/tclListObj.c
@@ -100,8 +100,8 @@ NewListIntRep(
 	    attemptckalloc(sizeof(List) + ((objc-1) * sizeof(Tcl_Obj *)));
     if (listRepPtr == NULL) {
 	if (p) {
-	    Tcl_Panic("list creation failed: unable to alloc %u bytes",
-		    sizeof(List) + ((objc-1) * sizeof(Tcl_Obj *)));
+	    Tcl_Panic("list creation failed: unable to alloc %lu bytes",
+		    (unsigned long) (sizeof(List) + ((objc-1) * sizeof(Tcl_Obj *))));
 	}
 	return NULL;
     }
@@ -163,8 +163,8 @@ AttemptNewList(
 		    LIST_MAX));
 	} else {
 	    Tcl_SetObjResult(interp, Tcl_ObjPrintf(
-		    "list creation failed: unable to alloc %u bytes",
-		    sizeof(List) + ((objc-1) * sizeof(Tcl_Obj *))));
+		    "list creation failed: unable to alloc %lu bytes",
+		    (unsigned long) (sizeof(List) + ((objc-1) * sizeof(Tcl_Obj *)))));
 	}
     }
     return listRepPtr;
